Dr_DocsGuru/requirements.txt
streamlit
faiss-cpu
numpy
scikit-learn
transformers
torch 